logout working, security updated, delete working
checks if cookies for teh remove and add favourite are null, checks poroperty object if its null, added prepeired statements to stop sql injection,delete working and confirm delete is working

+++ Assignment2/src/main/webapp/index.jsp
@@ -64,7 +64,7 @@
                     <td><a href="displayProperty?propertyId=${aProperty.id}"><img src="images/properties/thumbs/${aProperty.photo}"></td>
                             <shiro:user>
                                 <td><a href="editProperty?id=${aProperty.id}">Edit</a></td>
-                                <td><a href="deleteProperty?propertyId=${aProperty.id}">Delete</a></td>
+                                <td><a href="deleteProperty?propertyId=${aProperty.id}" onclick="return confirm('Are you sure?')">Delete</a></td>
                             </shiro:user>
                 </tr>
